back/pyproject.toml
[tool.poetry]
name = "back"
version = "0.1.0"
description = ""
authors = ["GauterHeart <gauterheart@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.77.1"
asyncpg = "^0.25.0"
redis = "^4.3.1"
aioredis = "^2.0.1"
python-dotenv = "^0.20.0"
uvicorn = "^0.17.6"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
